:msucle: only retain required values
Because we only care about the distance between the total two
strings (and not any substrings for example), we know we only need to
retain the previous row in order to produce the next row.

This gives a speedup of 10% for modest length strings (~7-15
characters), and a speedup of 17% for larger strings (64
characters). (On my macbook air)